pyproject.toml
[project]
name = "codex-localstack"
version = "0.1.0"
description = "Windows-first lazy-loaded local AI MCP gateway for Codex"
readme = "README.md"
requires-python = ">=3.11,<3.12"
dependencies = [
  "python-dotenv>=1.1.0",
  "httpx>=0.28.1",
  "mcp>=1.10.1",
  "lancedb==0.25.3",
  "fastembed>=0.7.3",
  "pymupdf>=1.26.0",
  "rapidocr-onnxruntime>=1.4.4",
  "pdfplumber>=0.11.6",
  "pillow>=11.3.0",
  "numpy>=2.3.0",
  "orjson>=3.11.0",
  "dulwich>=0.24.5",
  "pyarrow>=20.0.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.4.0",
]

[dependency-groups]
dev = [
  "pytest>=8.4.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]
